src/requirements.txt
dj-database-url==2.1.0
Django==4.2
django-allauth==0.56.1
django-debug-toolbar==4.1.0
django-hosts==5.2
django-rq==2.8.1
django-vite==2.1.3
pylint==2.17.5
python-dateutil==2.8.2
pytz==2023.3
redis==4.6.0
rq==1.15.1
stripe==6.1.0
uWSGI==2.0.21
